refactor(topology): isolate homology contracts
